+++ documentation/Changelog.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Dispatch barcode detector offset calculation
+* Fixed button appearance
 
 ### Updated
 * Updated localizations